separated join command into cpp file. Also cleaned out some comments
